pyproject.toml
[project]
name = "testingapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Wild",email = "zoppoz4100@gmail.com"}
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-playwright (>=0.7.2,<0.8.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "allure-pytest (>=2.15.3,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest-xdist (>=3.8.0,<4.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "ruff (>=0.15.6,<0.16.0)"
]
